1.2.0: Users can now deactivate their own accounts.

+++ models/User.php
@@ -1,13 +1,14 @@
 <?php namespace RainLab\User\Models;
 
-use URL;
 use Mail;
-use DB;
+use Event;
 use October\Rain\Auth\Models\User as UserBase;
 use RainLab\User\Models\Settings as UserSettings;
 
 class User extends UserBase
 {
+    use \October\Rain\Database\Traits\SoftDeleting;
+
     /**
      * @var string The database table used by the model.
      */
@@ -83,13 +84,37 @@ class User extends UserBase
         }
     }
 
+    public function afterLogin()
+    {
+        if ($this->trashed()) {
+            $this->last_login = $this->freshTimestamp();
+            $this->restore();
+
+            Mail::sendTo($this, 'rainlab.user::mail.reactivate', [
+                'name' => $this->name
+            ]);
+
+            Event::fire('rainlab.user.reactivate', [$this]);
+        }
+        else {
+            parent::afterLogin();
+        }
+    }
+
     /**
      * After delete event
      * @return void
      */
     public function afterDelete()
     {
+        if ($this->isSoftDelete()) {
+            Event::fire('rainlab.user.deactivate', [$this]);
+            return;
+        }
+
         $this->avatar && $this->avatar->delete();
+
+        parent::afterDelete();
     }
 
     public function scopeIsActivated($query)
@@ -156,14 +181,10 @@ class User extends UserBase
         }
 
         if ($mailTemplate = UserSettings::get('welcome_template')) {
-            $data = [
+            Mail::sendTo($this, $mailTemplate, [
                 'name'  => $this->name,
                 'email' => $this->email
-            ];
-
-            Mail::send($mailTemplate, $data, function($message) {
-                $message->to($this->email, $this->name);
-            });
+            ]);
         }
 
         return true;
